+++ src/autocomplete-options.ts
@@ -119,6 +119,12 @@ export default class AutocompleteOptions {
     wrapperClassName: string;
 
     /**
+     * Automatically clear the screen reader announcement element after the specified delay
+     * Defaults to 2 seconds if true
+     */
+    srAutoClear: boolean | number = false;
+
+    /**
      * Screen reader text used in multiple mode for element deletion.
      * Prepended to option label in aria-label attribute e.g. 'delete Canada'
      */