+++ src/autocomplete.ts
@@ -42,7 +42,8 @@ export default class Autocomplete {
     input: HTMLInputElement;
     wrapper: HTMLDivElement;
     showAll: HTMLSpanElement;
-    srAnnouncements: HTMLSpanElement;
+    srAssistance: HTMLParagraphElement;
+    srAnnouncements: HTMLParagraphElement;
 
     // non elements
     api: AutocompleteApi;
@@ -293,7 +294,7 @@ export default class Autocomplete {
      */
     isSelectedElem(element: Element): boolean {
         const sourceEntry = element && element[SELECTED_OPTION_PROP];
-        return this.multiple && !!sourceEntry && typeof sourceEntry === 'object';
+        return !!(this.multiple && typeof sourceEntry === 'object');
     }
 
     /**
@@ -426,7 +427,7 @@ export default class Autocomplete {
         // insert new elements
         // can't check against fragment.children or fragment.childElementCount, as does not work in IE
         if (fragment.childNodes && fragment.childNodes.length) {
-            this.wrapper.insertBefore(fragment, this.list);
+            this.wrapper.insertBefore(fragment, this.srAssistance);
         }
 
         // set ids on selected DOM elements
@@ -948,8 +949,8 @@ export default class Autocomplete {
             if (
                 !force &&
                 activeElem &&
-                !(this.showAll && this.showAll === activeElem) && // exception for show all button
-                !this.isSelectedElem(activeElem) && // exception for selected items
+                // exception for selected items, as these sit below the input by default
+                !this.isSelectedElem(activeElem) &&
                 // must base this on the wrapper to allow scrolling the list in IE
                 this.wrapper.contains(activeElem)
             ) {
@@ -1584,7 +1585,7 @@ export default class Autocomplete {
         clearTimeout(this.elementChangeEventTimer);
 
         // clear stored element vars
-        ['element', 'label', 'list', 'input', 'wrapper', 'showAll', 'srAnnouncements'].forEach(
+        ['element', 'label', 'list', 'input', 'wrapper', 'showAll', 'srAssistance', 'srAnnouncements'].forEach(
             (entry: string) => (this[entry] = null)
         );
     }
@@ -1621,7 +1622,8 @@ export default class Autocomplete {
         this.input = document.getElementById(this.ids.INPUT) as HTMLInputElement;
         this.wrapper = document.getElementById(this.ids.WRAPPER) as HTMLDivElement;
         this.showAll = document.getElementById(this.ids.BUTTON) as HTMLSpanElement;
-        this.srAnnouncements = document.getElementById(this.ids.SR_ANNOUNCEMENTS) as HTMLSpanElement;
+        this.srAssistance = document.getElementById(this.ids.SR_ASSISTANCE) as HTMLParagraphElement;
+        this.srAnnouncements = document.getElementById(this.ids.SR_ANNOUNCEMENTS) as HTMLParagraphElement;
 
         // set internal source array, from static elements if necessary
         this.prepListSource();