+++ src/autocomplete-options.ts
@@ -75,6 +75,17 @@ export default class AutocompleteOptions {
     deleteOnBackspace: boolean = false;
 
     /**
+     * In multiple mode, if more than 1 item is selected,
+     * add a button at the beginning of the selected items as a shortcut to delete all
+     */
+    deleteAllControl: boolean = false;
+
+    /**
+     * Text to use in the deleteAllControl
+     */
+    deleteAllText: string = `Delete all`;
+
+    /**
      * In async mode, parameter to use when adding the input value to the
      * endpoint String. e.g. https://some-endpoint?q=norway&limit=9999
      */